Description of sizes

+++ 0x00-hello_world/6-size.c
@@ -13,11 +13,11 @@ int main(void)
 	char c;
 	float f;
 
-	printf("Size of a char: %lu byte(s)\n", (sizeof(c)));
-	printf("size of an int: %lu byte(s)\n", (sizeof(i)));
-	printf("Size of a long int: %lu byte(s)\n", (sizeof(l)));
-	printf("Size of a long long int: %lu byte(s)\n", (sizeof(j)));
-	printf("Size of a float: %lu byte(s)\n", (sizeof(f)));
+	printf("Size of a char: %lu byte(s)\n",(unsigned long)(sizeof(c)));
+	printf("size of an int: %lu byte(s)\n", (unsigned long)(sizeof(i)));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)(sizeof(l)));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)(sizeof(j)));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)(sizeof(f)));
 
 	return (0);
 }
